requirement.txt
Django==1.11.6
django-allauth==0.33.0
django-bootstrap-form==3.4
django-crispy-forms==1.7.2
django-filter==1.1.0
django-filters==0.2.1
django-markdown-deux==1.0.5
django-pagedown==1.0.6
django-widget-tweaks==1.4.3
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
docopt==0.6.2
idna==2.6
Markdown==2.6.9
markdown2==2.3.6
oauthlib==2.0.6
Pillow==5.2.0
pipreqs==0.4.9
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
xlwt==1.3.0
yarg==0.1.9
gunicorn
django-heroku